main_backend_service/requirements.txt
Flask
youtube-transcript-api 
openai
python-dotenv
langchain 
langchain-openai
langchain-core
langchain-community
flask-cors
flask_socketio
redis
PyMuPDF
pillow
py-mon
supabase
gunicorn
gevent-websocket
aiohttp
asyncio
langchain_openai
pytube 
google-api-python-client

beautifulsoup4
requests
